+++ addon/prefs.js
@@ -1,1 +1,2 @@
 pref("cleanChapterNumbers", true);
+pref("interfaceLanguage", "zh-CN");